requirements.txt
antlr4-python3-runtime==4.9.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
